strip Tags from analysis
